+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.8.0] - 2023-12-22
+### Added
+- `countEntities()` method and corresponding `count()` route handler.
+This allows counting the rows that match a query/filter without selecting them.
+
 ## [2.7.0] - 2023-09-27
 ### Added
 - `writableId` bool property to optionally make the ID column writable.
@@ -146,7 +151,8 @@ return early if passed an empty IDs array.
 ### Changed
 - Initial stable release
 
-[Unreleased]: https://github.com/theodorejb/phaster/compare/v2.7.0...HEAD
+[Unreleased]: https://github.com/theodorejb/phaster/compare/v2.8.0...HEAD
+[2.8.0]: https://github.com/theodorejb/phaster/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/theodorejb/phaster/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/theodorejb/phaster/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/theodorejb/phaster/compare/v2.4.0...v2.5.0